Remove bone controller DAO and separate child interfaces.
