+++ feature_bone/src/main/java/com/hfad/palamarchuksuperapp/feature/bone/data/local/dao/GeneralDao.kt
@@ -1,0 +1,5 @@
+package com.hfad.palamarchuksuperapp.feature.bone.data.local.dao
+
+interface GeneralDao {
+    suspend fun clearAllTables()
+}
